fix bug in log_reg training
git-svn-id: file:///home/gumz3/Desktop/svnAugustusRepo/svnArchiveFile/svn/augustus/trunks@1302 946bced8-b911-4d30-a262-e0a769a8c5ad

Former-commit-id: 838a73e66a3a4ed949b3d03f0576b4dc235493e0

+++ include/train_logReg_param.hh
@@ -34,8 +34,8 @@ public:
   vector<pair<int, vector<double> > > *exon_samples;
   vector<pair<int, vector<double> > > *intron_samples;
   int num_features;
-  int prob_true_false_exons;
-  int prob_true_false_introns;
+  double prob_true_false_exons;
+  double prob_true_false_introns;
   vector<int> exon_feature_idx;
   vector<int> intron_feature_idx;
   void set_mean_std(vector<pair<int, vector<double> > > *samples, vector<int> feature_idx);
